Break out Camera components into separate files
