+++ client/camera/CameraMain.css
@@ -1,28 +1,3 @@
-:local(.knob) {
-  @value size: 10px;
-  position: absolute;
-  margin-left: calc(size / -2);
-  margin-top: calc(size / -2);
-  width: size;
-  height: size;
-  box-shadow: 0 0 0 1px red;
-  border-radius: 1000px;
-  cursor: pointer;
-}
-:local(.knob):hover {
-  background: rgba(255, 0, 0, 0.5);
-}
-
-:local(.keyPoint) {
-  position: absolute;
-  background: rgba(255, 255, 255, 0.3);
-  border-radius: 1000px;
-  cursor: pointer;
-}
-:local(.keyPoint):hover {
-  background: rgba(255, 255, 255, 0.6);
-}
-
 @value cameraMainPadding: 20px;
 @value cameraMainSidebarWidth: 300px;
