+++ client/camera/Knob.css
@@ -1,0 +1,14 @@
+:local(.knob) {
+    @value size: 10px;
+    position: absolute;
+    margin-left: calc(size / -2);
+    margin-top: calc(size / -2);
+    width: size;
+    height: size;
+    box-shadow: 0 0 0 1px red;
+    border-radius: 1000px;
+    cursor: pointer;
+  }
+  :local(.knob):hover {
+    background: rgba(255, 0, 0, 0.5);
+  }
